fix util library and add some comments
