+++ src/lib/customerror/node-customerror.d.ts
@@ -1,4 +1,6 @@
+/**
+ * base class for creating custom error class
+ */
 export declare class BaseCustomError extends Error{
   constructor(message?: string);
-  protected initializeError(c: Function): void;
 }
