+++ src/lib/util/util.d.ts
@@ -1,5 +1,12 @@
+/**
+ * clone util function
+ *
+ * @param obj source object
+ * @param deepcopy is deep copying?
+ * @returns clone object of source
+ */
 export declare function clone<T>(obj: T, deepcopy?: boolean): T;
+
 export declare function cloneObj<T>(obj: T): T;
+
 export declare function inherits<T, U>(ctor: T, superCtor: U): T & U;
-export declare function repeat(count: number, str: string): string;
-export declare function escapeRegExp(str: string): string;
