+++ src/lib/util/util.js
@@ -53,17 +53,3 @@ var inherits = module.exports.inherits = function(ctor, superCtor) {
     }
   });
 };
-
-
-var repeat = module.exports.repeat = function(count, str) {
-  var repU;
-  if (count <= 0) {
-    return "";
-  }
-  repU = repeat(Math.floor(count / 2), str);
-  return repU + repU + ((count % 2) == 1 ? str : "");
-}
-
-var escapeRegExp = module.exports.escapeRegExp = function(str) {
-  return str.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
