+++ src/parseresult/parseerr.ts
@@ -1,11 +1,23 @@
+/**
+ * Node type for Parse Error children
+ */
 export type ParseErrorNode = [string, ParseErrorStocker];
 
+/**
+ * Error stocker class
+ */
 export class ParseErrorStocker {
   private innerMsg: string;
   private innerExpected: string;
   private innerActual: string;
   private innerChilds: ParseErrorNode[];
 
+  /**
+   * @param childs children of stocker
+   * @param msg failure message
+   * @param exp expected type
+   * @param act actual object show
+   */
   constructor(childs?: ParseErrorNode[]);
   constructor(msg: string, childs?: ParseErrorNode[]);
   constructor(msg: string, exp?: string, childs?: ParseErrorNode[]);
@@ -36,7 +48,20 @@ export class ParseErrorStocker {
     }
   }
 
+  /**
+   * return new stocker added child
+   *
+   * @param value added child value
+   * @returns new stocker including new child
+   */
   addChild(value: ParseErrorNode): ParseErrorStocker;
+  /**
+   * return new stocker added child
+   *
+   * @param pname property name
+   * @param stocker child stocker value
+   * @returns new stocker including new child
+   */
   addChild(pname: string, stocker: ParseErrorStocker): ParseErrorStocker;
   addChild(arg1: (string|ParseErrorNode), arg2?: ParseErrorStocker) {
     if (typeof arg1 === "string") {
@@ -46,14 +71,32 @@ export class ParseErrorStocker {
     }
   }
 
+  /**
+   * return new stocker added children
+   *
+   * @param nodes nodes for adding
+   * @returns new stocker including new children
+   */
   addChilds(nodes: ParseErrorNode[]) {
     return new ParseErrorStocker(this.innerMsg, this.innerExpected, this.innerChilds.concat(nodes));
   }
 
+  /**
+   * return new stocker redesced
+   *
+   * @param msg new failure message
+   * @param exp new expected type
+   * @returns new stocker redescing
+   */
   desc(msg: string, exp?: string) {
     return new ParseErrorStocker(msg, (typeof exp === "undefined") ? this.innerExpected : exp, this.innerChilds);
   }
 
+  /**
+   * report this error with report function
+   *
+   * @param f report function
+   */
   report(f: (msg: string, exp?: string, act?: string, childs?: ParseErrorNode[]) => any): void {
     f(this.innerMsg, this.innerExpected, this.innerActual, this.innerChilds);
   }
